Extract form utilities
